Fix bug that added extra dot to file extension
The various `*_extension` variables are meant to take the dot into
account (i.e. a valid extension should be, for example `".md"`).  The
regex that was looking for a match was adding an extra dot, causing the
match to fail if the file already had the extension.  This effectively
deals with the issue of the file extension being present or not in the
link text.